[Change] move theme to utils
